Type: Including a description of the algorithm
Description: I included a description of the OpenCV library about the
Viola and Jones work about face detection. The aim of the notebook is to
introduce the computer vision task of detecting faces on an image. For
that reason was important to describe on a high level the Haar
feature-like algorithm used